moved search implementation into registry package, return github packages data instead of table regs
